+++ df-logger/CHANGELOG.md
@@ -1,6 +1,7 @@
 # DF User Logger
 
 ## Release 1.6.6 (2022-01-04)
+- **NEW::** Better default message management for any number of localized messages.
 - **FIX #281:** Sounds folder was missing for previous release.
 
 ## Release 1.6.5 (2022-01-02)